requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
impyla>=0.18.0
psycopg2-binary>=2.9.0
optuna>=3.0.0
catboost>=0.26.0
openpyxl>=3.0.0
pyarrow>=7.0.0
pytest>=8.0.0
python-docx>=0.8.11
matplotlib>=3.5.0
seaborn>=0.11.2 